requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
akanekopy==1.1.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
cloudscraper==1.2.60
config==0.5.1
discord-components==2.1.2
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
jsons==1.6.1
multidict==6.0.2
psutil==5.9.0
pyfiglet==0.8.post1
pyparsing==3.0.8
python-dateutil==1.5
python-utils==3.1.0
requests==2.27.1
requests-toolbelt==0.9.1
soupsieve==2.3.2.post1
typing-extensions==3.10.0.2
typish==1.9.3
urllib3==1.26.9
utils==1.0.1
xmltodict==0.12.0
yarl==1.7.2